requirements.txt
dnspython==2.6.1
paho-mqtt==1.6.1
pymongo==4.8.0
pytz==2022.1
PyYAML==6.0.1
annotated-types==0.7.0
anyio==3.7.1
click==8.1.7
colorama==0.4.6
fastapi==0.103.2
h11==0.14.0
idna==3.7
pydantic==2.8.2
pydantic_core==2.20.1
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.12.2
uvicorn==0.23.2